requirements.txt
streamlitstreamlit
pandas
numpy